cleanup warnings
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@4720 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_zgesdd.cpp
@@ -180,7 +180,7 @@ int main( int argc, char** argv)
                 }
                 
                 default: {
-                    fprintf( stderr, "Error: unknown option svd_work %d\n", opts.svd_work );
+                    fprintf( stderr, "Error: unknown option svd_work %d\n", (int) opts.svd_work );
                     exit(1);
                     break;
                 }
